requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
openai==1.3.7
playwright==1.40.0
redis==5.0.1
aioredis==2.0.1
pandas==2.1.4
numpy==1.25.2
fuzzywuzzy[speedup]==0.18.0
python-dotenv==1.0.0
structlog==23.1.0
loguru==0.7.0
pytest==7.4.3
pytest-asyncio==0.21.1
rich 
streamlit==1.32.2
streamlit-folium==0.18.0
folium==0.16.0
plotly==5.21.0 